Use citeproc-js for hard-coded citations.
