+++ website/static/js/pages/project-dashboard-page.js
@@ -16,7 +16,8 @@ var Raven = require('raven-js');
 
 var NodeControl = require('../nodeControl.js');
 
-var CitationWidget = require('../citations.js');
+var CitationList = require('../citationList.js');
+var CitationWidget = require('../citationWidget.js');
 
 var ctx = window.contextVars;
 var nodeApiUrl = ctx.node.urls.api;
@@ -29,9 +30,7 @@ $('body').on('nodeLoad', function(event, data) {
     new LogFeed('#logScope', nodeApiUrl + 'log/');
     // Initialize nodeControl
     new NodeControl('#projectScope', data);
-
 });
-
 
 // Initialize comment pane w/ it's viewmodel
 var $comments = $('#comments');
@@ -44,6 +43,7 @@ if ($comments.length) {
 
 // Initialize CitationWidget if user isn't viewing through an anonymized VOL
 if (!ctx.node.anonymous) {
+    new CitationList('#citationList');
     new CitationWidget('#citationStyleInput', '#citationText');
 }
